Added output file and parameter file, fixed minor bugs
The PHONON.out file contains one line information about the run. It will not be overwritten by each run, it is meant to ba able to plot the results.

The PARAM.out contains all the parameter of the run that created its parent folder.

Fixed bug: issues a warning when the defect concentration is to low. (instead of falling)