Arbitrarly sized defects implementation [NOT OPTIMAL]
The way to enter defects is now more permissive, the code will raise an error if the mass vector and the potential don't match.
If the SC size is not a multiple of the PC size--which is now possible, the code finds the lcm and makes the SC bigger. This is not optimal, as there must be a way to have periodicity in the Super Cell that would avoid solving a bigger SC. This is a minor issue since solving the SC matrix takes very small portion of the total computation time.