PhononCode as a module + MonteCarlo Optimizer
The PhononCode can now be called as function in addition to the usual argument call, the function can be call with or without specifying the defect vector. The display and output can be completely turned of to get only the LTC.

The MonteCarlo optimizes the LTC with respect to the defect layout.